backend/requirements.txt
flask
flask-cors
openai
numpy
faiss-cpu
sentence-transformers
scikit-learn
googletrans
langdetect